Fix door pass-through

+++ scripts/classes/door.gd
@@ -3,8 +3,11 @@ extends "res://scripts/classes/prop.gd"
 
 const Hero = preload("res://scripts/classes/hero.gd")
 
+var _open = false
+
 onready var lock = get_node("lock")
 onready var anim = get_node("lock/AnimationPlayer")
+onready var teleport = get_node("teleport")
 
 export(int) var target_stage = 0
 export(String) var target_door = "door"
@@ -13,46 +16,42 @@ export(bool) var start_opened = true
 
 signal change_stage(which, door)
 
-var locked = false
-
 func _ready():
-  if start_opened:
+  if start_opened or _open:
     open()
-  else:
+  elif !_open:
     close()
-  if key_required != null:
-    locked = true
-  else:
-    locked = false
-  get_node("teleport").set_layer_mask(_mask)
-  get_node("teleport").set_collision_mask(_mask)
+  printt("mask", _mask)
+  printt("open", _open)
+  teleport.set_layer_mask(_mask)
+  teleport.set_collision_mask(_mask)
 
 func open():
-  if anim.get_current_animation() != "opened":
-    print("door open")
-    anim.play("opening")
-    anim.queue("opened")
-    deactivate_collision()
+  print("door open")
+  anim.play("opening")
+  anim.queue("opened")
+  deactivate_collision()
+  _open = true
+  set_fixed_process(true)
 
 func close():
-  if anim.get_current_animation() != "closed":
-    print("door closed")
-    anim.play("closing")
-    anim.queue("closed")
-    activate_collision()
+  print("door closed")
+  anim.play("closing")
+  anim.queue("closed")
+  activate_collision()
+  _open = false
+  set_fixed_process(false)
 
 func is_closed():
   return anim.get_current_animation() != "opened"
 
 func unlock(key):
   if key != null and key.get_name() == key_required:
-    locked = false
+    open()
     return true
   return false
 
 func on():
-  if key_required != null and locked:
-    return
   open()
 
 func get_spawn_pos():
@@ -61,14 +60,22 @@ func get_spawn_pos():
 func off():
   close()
 
+func _fixed_process(dt):
+  var bodies = teleport.get_overlapping_bodies()
+  if bodies.size() > 0:
+    print("tp bodies= ", bodies)
+    for body in bodies:
+      if body.has_method("get_layer_mask") and body.get_layer_mask() == teleport.get_collision_mask():
+        _teleport(body)
+
 func _teleport(body):
+  printt(get_path(), "teleport!", body.get_path())
   if is_inside_tree() and body.get_script() == Hero and anim.get_current_animation() == "opened":
     printt(get_path(), "door used!", target_stage, target_door)
     get_node("/root/main/gameplay")._change_stage(target_stage, target_door)
 
 func interact(body):
   #emit_signal("change_stage", target_stage,  target_door)
-  if locked:
-    if body.get_pocket_item() != null:
-      if unlock(body.get_pocket_item()):
-        body.drop()
+  if is_closed() and body.get_pocket_item() != null:
+    if unlock(body.get_pocket_item()):
+      body.drop()
